requirements.txt
blinker==1.6.2
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-decouple==3.8
SQLAlchemy==2.0.13
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.4
